requirements.txt
connexion[swagger-ui]
flask_sqlalchemy
transitions
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
